v1.0.4 - Modlist and general fixes.
- Moderator list added with -=mods. Get a full list with -=mods or check if theyre a mod with -=checkmod.
- Fixed retrieving icons from the Icon API. Now if there is an error, it says so.
- Fix in help so now it doesnt error when sending a help message to a DM

+++ commands/icon.js
@@ -15,6 +15,7 @@ module.exports.run = async (bot, message, args) => {
 	profile.cube = profile.icon;
 	if (a === 2) var img = await gdtools.createIcon(profile, args[1]);
 	else var img = await gdtools.createIcons(profile);
+	if (img === 'error') return msg.edit(`There was an issue getting these images. Perhaps the Icon API is down.`);
 	msg.delete();
 	message.channel.send(
 		new Discord.RichEmbed()
